requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.36.18
botocore==1.36.18
click==8.1.8
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
numpy==2.2.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
